Another set of code review comment fixes
- move everything to single InterimCanvasPaintingX
- sort out spacing in ManagedAssetResultFinder
- additional tests on JObjectX
- simplify some logic in CanvasPaintingMerger
