+++ src/IIIFPresentation/API/Features/Manifest/ManagedAssetResultFinder.cs
@@ -143,13 +143,14 @@ public class ManagedAssetResultFinder(
 
         var dlcsAssetIds = dlcsAssets.Select(d => d.GetAssetId(customerId));
 
-        var missingAssets = interimCanvasPaintings.Where(icp =>
-                !dlcsAssetIds.Any(a => icp.SuspectedAssetId == a.Asset && icp.CustomerId == a.Customer && icp.SuspectedSpace == a.Space))
+        var missingAssets = interimCanvasPaintings.Where(icp => !dlcsAssetIds.Any(a =>
+                icp.SuspectedAssetId == a.Asset && icp.CustomerId == a.Customer && icp.SuspectedSpace == a.Space))
             .Select(icp => icp.Id).ToList();
         
         if (missingAssets.Count != 0)
         {
-            throw new PresentationException($"Managed assets not found: {string.Join(", ", missingAssets.Select(a=>$"'{a}'"))}");
+            throw new PresentationException(
+                $"Managed assets not found: {string.Join(", ", missingAssets.Select(a => $"'{a}'"))}");
         }
     }
